+++ packages/docs/src/data/api/notification-module.ts
@@ -1,0 +1,30 @@
+import { Api } from '~/types';
+
+export const api: Api = {
+	addNotification: {
+		arguments: [
+			{
+				name: 'notification',
+				description: 'La notification à afficher.',
+				type: 'NotificationObj',
+				required: true,
+				example: `type: 'success' | 'error' | 'info' | 'warning';
+message: string;
+icon?: string;
+ref?: string;`
+			}
+		],
+		returnValue: [
+			{
+				type: 'void'
+			}
+		]
+	},
+	clearNotification: {
+		returnValue: [
+			{
+				type: 'void'
+			}
+		]
+	}
+};